BACKEND/geosense-server/med_chatbot/requirements.txt
python-dotenv
langchain
pinecone-client
sentence-transformers
groq
langchain-community
langchain_groq
langchain_pinecone
langchain_huggingface
pypdf